handle missing REDMINE_API_KEY
Prior to this commit, when REDMINE_API_KEY was not set in a
configuration, the plugin would crash with an AttributeError (rather
than the NameError that is caught.) This caused the test suite to fail.

Instead of trying to catch the specific error when REDMINE_API_KEY is
not set, update the REDMINE_API_KEY conditional to use hasattr()
instead.

Signed-off-by: Ken Dreyer <ktdreyer@ktdreyer.com>