make HTTP calls async with Twisted and treq
Prior to this change, the requests.get() call would block Helga's
Twisted reactor from responding to any other events. In practicality,
this meant that if I triggered the Redmine plugin in my channel, and
then immediately tried another Helga command:

  < ktdreyer> I'm working on issue 1234
  < ktdreyer> helga: ping

... then Helga would not process or respond to the "ping" command or any
other inputs until the HTTP request for ticket 1234 completed (or timed
out, etc).

  [... after HTTP request to Redmine completes ...]
  < helga> ktdreyer might be talking about
           http://tracker.example.com/issues/1234
  < helga> pong

Use the Twisted version of requests, "treq", send the GET request and
return control to the Twisted reactor to await the HTTP response. Move
the HTTP processing and subsequent reply message ("foo might be talking
about bar") to callback functions that Twisted will call when it
receives the HTTP response from the Redmine web server.

This keeps Helga from hanging if Redmine is offline or slow.

(treq 15.1.0 is the first version to include StubTreq, so we set that
minimum number in setup.py)