MANIFEST.in: include redmine.py in packaging
Prior to this change, setuptools would not include redmine.py in the
sdist tarball.

This would lead to errors when trying to install helga-redmine directly
from pypi:

  ImportError: No module named redmine

The setuptools find_package() call is ineffective and can probably be
removed (unless we want to move redmine.py to
helga_redmine/__init__.py).